requirements.txt
Flask==3.0.0
zeep==4.2.1
lxml==5.1.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0